Move exceptions to their own package and subclass one class.
