+++ core/src/main/java/org/nwapw/abacus/exception/AbacusException.java
@@ -1,0 +1,9 @@
+package org.nwapw.abacus.exception;
+
+public class AbacusException extends RuntimeException {
+
+    public AbacusException(String baseMessage, String description){
+        super(baseMessage + ((description == null) ? "." : (": " + description)));
+    }
+
+}
