+++ core/src/main/java/org/nwapw/abacus/exception/DomainException.java
@@ -1,24 +1,24 @@
-package org.nwapw.abacus.function;
+package org.nwapw.abacus.exception;
 
 /**
  * Exception thrown if the function parameters do not match
  * requirements.
  */
-public class DomainException extends RuntimeException {
+public class DomainException extends AbacusException {
 
     /**
      * Creates a new DomainException.
      * @param reason the reason for which the exception is thrown.
      */
     public DomainException(String reason) {
-        super(reason);
+        super("Domain error", reason);
     }
 
     /**
      * Creates a new DomainException with a default message.
      */
     public DomainException(){
-        this("Domain error.");
+        this(null);
     }
 
 }
