+++ core/src/main/java/org/nwapw/abacus/exception/EvaluationException.java
@@ -1,17 +1,17 @@
-package org.nwapw.abacus.function;
+package org.nwapw.abacus.exception;
 
 /**
  * An exception thrown primarily from Tree Value operators and functions,
  * which have to deal with the result of a Reducer as well as the results
  * of Applicable.
  */
-public class EvaluationException extends RuntimeException {
+public class EvaluationException extends AbacusException {
 
     /**
      * Creates a new EvaluationException with the default string.
      */
     public EvaluationException() {
-        super("Evaluation error.");
+        this(null);
     }
 
     /**
@@ -19,7 +19,7 @@ public class EvaluationException extends RuntimeException {
      * @param message the message string.
      */
     public EvaluationException(String message) {
-        super(message);
+        super("Evaluation error", message);
     }
 
 }
