+++ fx/src/main/java/org/nwapw/abacus/fx/AbacusController.java
@@ -12,10 +12,11 @@ import javafx.util.Callback;
 import javafx.util.StringConverter;
 import org.nwapw.abacus.Abacus;
 import org.nwapw.abacus.config.Configuration;
+import org.nwapw.abacus.exception.ComputationInterruptedException;
 import org.nwapw.abacus.function.Documentation;
 import org.nwapw.abacus.function.DocumentationType;
-import org.nwapw.abacus.function.DomainException;
-import org.nwapw.abacus.function.EvaluationException;
+import org.nwapw.abacus.exception.DomainException;
+import org.nwapw.abacus.exception.EvaluationException;
 import org.nwapw.abacus.number.*;
 import org.nwapw.abacus.plugin.ClassFinder;
 import org.nwapw.abacus.plugin.PluginListener;
